move auth provider to navbar
